+++ AOS_BE/src/main/java/com/aos/AOSBE/Entity/PromotionProduct.java
@@ -23,7 +23,7 @@ public class PromotionProduct {
     @Column(name = "is_gift")
     private boolean isGift;
     @Column(name = "cost_share")
-    private int costShare;
+    private double costShare;
     @Column(name = "created_at")
     private LocalDateTime createdAt;
     @Column(name = "updated_at")
